+++ pkg/providers/reportinggroups/provider_test.go
@@ -1,0 +1,11 @@
+package reportinggroups
+
+import (
+	"testing"
+
+	"github.com/akamai/terraform-provider-akamai/v10/pkg/common/testutils"
+)
+
+func TestMain(m *testing.M) {
+	testutils.TestRunner(m)
+}