+++ pkg/providers/reportinggroups/reportinggroups.go
@@ -1,0 +1,7 @@
+package reportinggroups
+
+import "github.com/akamai/terraform-provider-akamai/v10/pkg/providers/registry"
+
+func init() {
+	registry.RegisterSubprovider(NewSubprovider())
+}