+++ pkg/providers/reportinggroups/testdata/TestDataCPCode/cp_code.tf
@@ -1,0 +1,8 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_reportinggroups_cp_code" "test" {
+  cp_code_id = 12345
+}
+